order page

+++ backendEcom/src/main/java/com/niit/models/Product.java
@@ -35,7 +35,7 @@ public class Product implements Serializable {
  
 @Column
 @Min(value=1)
- double qty;
+ int qty;
  @Column
  @Min(value=1)
  double price;
@@ -100,10 +100,10 @@ public int getProductid() {
 	public void setImagePath(String imagePath) {
 		this.imagePath = imagePath;
 	}
-public double getQty() {
+public int getQty() {
   return qty;
  }
- public void setQty(double qty) {
+ public void setQty(int qty) {
   this.qty = qty;
  }
  public double getPrice() {
